requirements.txt
# Core AI
anthropic>=0.40.0

# Web Server (dashboard + mobile API)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Exchange & HTTP
requests>=2.31.0
websocket-client>=1.7.0
aiohttp>=3.9.0

# Data & Analysis
pandas>=2.1.0
numpy>=1.26.0
ta>=0.11.0

# ML / AI
scikit-learn>=1.4.0
joblib>=1.3.0

# News & Feeds
feedparser>=6.0.0
python-dateutil>=2.8.2

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Scheduling & UI
schedule>=1.2.0
colorlog>=6.8.0
rich>=13.7.0

# Security & Authentication
pyotp>=2.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
qrcode>=7.4.2
